requirements.txt
streamlit
pytesseract
opencv-python-headless
pandas
pillow
pdf2image
